Step 3

+++ includes/api-server.php
@@ -15,10 +15,10 @@ class API_Server {
 		$this->root_directory = $root_directory;
 		$this->db = new Database( $this->root_directory );
 
-		$this->match_path();
+		$this->parse_endpoint();
 	}
 
-	protected function match_path() {
+	protected function parse_endpoint() {
 
 		$parts = explode( '/', $this->request_uri );
 
@@ -29,19 +29,40 @@ class API_Server {
 
 	protected function pundits_list() {
 		$this->db->fetchAll('pundits');
-		$this->response = $this->db->get_last_result();
+		$this->response = array(
+			'status' => 'success',
+			'data' => $this->db->get_last_result()
+		);
 	}
 
 	protected function pundits_delete() {
+
+		if ( ! isset($_POST['id'] ) ) {
+
+			$this->response = array(
+				'status'    => 'error',
+				'code'      => 'missing-parameter',
+				'message'   => 'No pundit ID was provided.'
+			);
+
+			return;
+		}
 
 		// do some safety shit.
 		$id = $_POST['id'];
 
 		$this->db->delete('pundits', $id);
+
+		$this->response = array(
+			'status'    => 'success'
+		);
 	}
 
 	protected function pundits_update() {
 		// do some safety shit.
+
+
+
 		$id = $_POST['id'];
 		$data = $_POST['data'];
 		$this->db->update('pundits', $id, $data);
@@ -60,6 +81,15 @@ class API_Server {
 
 	public function run() {
 
+		if ( ! in_array( $_SERVER['REQUEST_METHOD'], array('GET', 'POST') ) ) {
+			$this->response = array(
+				'status'    => 'error',
+				'code'      => 'invalid-method',
+				'message'   => 'Invalid HTTP Method.'
+			);
+			return;
+		}
+
 		switch ( $this->endpoint ) {
 			case 'list':
 				$this->pundits_list();
@@ -71,7 +101,7 @@ class API_Server {
 				$this->pundits_delete();
 				break;
 			case 'create':
-				$this->pundits_list();
+//				$this->pundits_list();
 				break;
 			default:
 				$this->failure_response();
@@ -81,6 +111,8 @@ class API_Server {
 	}
 
 	public function output() {
+//		header("HTTP/1.0 404 Not Found"); TODO - http status codes?
+
 		header('Content-Type: application/json');
 		echo json_encode( $this->response );
 		exit;
